[SSN-102][ES] Refactor and place pagination as a separate component (reusable for other components).
